refactor(github-workflow): finish god-file decomposition (mod.rs 4445->190)
Extract the remaining 6 modules from the github_issue_workflow composition root: capability_profiles, approval_policy, stage_turn_submitter, stage_result_sink (+ its sink test mod), spawn (+ the provider-runtime + config-source test mods), completing the module tree begun earlier. mod.rs is now a 190-line thin root (shared consts + the workflow_invalid_config helper + the re-export wall). All production modules are well under 800 lines; test-only re-exports are cfg-gated to keep both the test-support and CLI feature combos warning-clean. Pure move: spawn_github_issue_workflow/GithubIssueWorkflowRuntimeDeps signatures unchanged; 923 composition tests pass.
